+++ src/Commands/RunExamples/CombineExampleOutput.php
@@ -1,0 +1,87 @@
+<?php
+
+/**
+ * Copyright (c) 2017-present Ganbaro Digital Ltd
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ *
+ *   * Redistributions of source code must retain the above copyright
+ *     notice, this list of conditions and the following disclaimer.
+ *
+ *   * Redistributions in binary form must reproduce the above copyright
+ *     notice, this list of conditions and the following disclaimer in
+ *     the documentation and/or other materials provided with the
+ *     distribution.
+ *
+ *   * Neither the names of the copyright holders nor the names of his
+ *     contributors may be used to endorse or promote products derived
+ *     from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+ * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+ * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+ * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+ * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ *
+ * @author    Stuart Herbert <stuherbert@ganbarodigital.com>
+ * @copyright 2017-present Ganbaro Digital Ltd www.ganbarodigital.com
+ * @license   http://www.opensource.org/licenses/bsd-license.php  BSD License
+ * @link      http://ganbarodigital.github.io/bengi
+ */
+
+namespace GanbaroDigital\Bengi\Commands\RunExamples;
+
+use GanbaroDigital\Bengi\Config;
+use GanbaroDigital\Bengi\Helpers;
+
+class CombineExampleOutput
+{
+    public static function using(array $capturedOutput)
+    {
+        $retval=[];
+        $mergedKeys=[];
+
+        foreach ($capturedOutput as $key1 => $output1) {
+            // avoid inception
+            if (isset($mergedKeys[$key1])) {
+                continue;
+            }
+
+            // we're going to want this in its own output block
+            $finalKey=[$key1];
+
+            // does anything else match?
+            foreach($capturedOutput as $key2 => $output2) {
+                // avoid inception
+                if ($key1 == $key2) {
+                    continue;
+                }
+
+                // are we already sorted?
+                if (isset($mergedKeys[$key2])) {
+                    continue;
+                }
+
+                if ($output1 === $output2) {
+                    $finalKey[] = $key2;
+                    $mergedKeys[$key2] = true;
+                }
+            }
+
+            $retval[implode(", ", $finalKey)] = $output1;
+        }
+
+        return $retval;
+    }
+}
+